requirements.txt
aniso8601==9.0.1
anyio==3.6.1
asgiref==3.5.2
backpack==0.1
backports.cached-property==1.0.1
blinker==1.4
cleo==0.6.8
click==8.1.3
Faker==0.8.18
fastapi==0.78.0
graphene==3.1
graphql-core==3.2.1
graphql-relay==3.2.0
h11==0.13.0
idna==3.3
inflection==0.3.1
lazy-object-proxy==1.7.1
orator==0.9.9
pastel==0.1.1
pendulum==1.5.1
promise==2.3
psycopg2-binary==2.9.3
pyaml==16.12.2
pydantic==1.9.1
Pygments==2.12.0
pylev==1.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-gql==0.2.4
python-multipart==0.0.5
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
Rx==1.6.1
simplejson==3.17.6
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
strawberry-graphql==0.114.0
text-unidecode==1.2
typing-extensions==4.2.0
tzlocal==1.5.1
uvicorn==0.17.6
wrapt==1.14.1
